refactor(desktop): extract updater env/config resolution into updater-config
Mechanical move of DESKTOP_UPDATE_ENV, the env validators,
resolveDesktopUpdaterConfig, and capabilitiesFor out of updater.ts into a
new updater-config module. updater.ts re-exports the public surface so
runtime wiring and existing imports are unchanged. The pure config test
cases move to updater-config.test.ts; behavior and total test count are
unchanged.
